major workaround to skip corrupt records in the binary file
HACK - When encountering corrupt records in the binary file
format we know replace the data with question marks and make
a best effort to skip ahead and continue instead of throwing
exceptions.  This allows us to recover the data and migrate
back to memory tables.  We can make a later attempt at restoring
the bad records from previous backups once the tables are in
a consistent state.